Move MovieFinderApplication to app root directory.
