+++ app/src/main/java/me/zwsmith/moviefinder/MoveFinderApplication.kt
@@ -1,7 +1,8 @@
-package me.zwsmith.moviefinder.core.dependencyInjection
+package me.zwsmith.moviefinder
 
 import android.app.Application
 import me.zwsmith.moviefinder.core.dependencyInjection.dagger.AppModule
+import me.zwsmith.moviefinder.core.dependencyInjection.dagger.DaggerInjector
 import me.zwsmith.moviefinder.core.dependencyInjection.dagger.Injector
 
 class MoveFinderApplication : Application() {
